add DiffcultyDisplay Comp
